% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdimg_set_path.R
\name{has_pdfimages}
\alias{has_pdfimages}
\title{Do you have pdfimages?}
\usage{
has_pdfimages()
}
\value{
boolean
}
\description{
Do you have pdfimages?
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
has_pdfimages()
\dontshow{\}) # examplesIf}
}
