% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdimg_detect.R
\name{pdimg_detect}
\alias{pdimg_detect}
\title{pdimg_detect}
\usage{
pdimg_detect(x)
}
\arguments{
\item{x}{output from \code{\link[=pdimg_images]{pdimg_images()}}}
}
\value{
xx
}
\description{
attempt to detect the "actual" plots in pdfs
}
\examples{
x1 <- system.file("examples/Tierney2017JOSS.pdf", package="pdfimager")
x2 <- system.file("examples/FuHughey2019.pdf", package="pdfimager")
x3 <- system.file("examples/vanGemert2018.pdf", package="pdfimager")
res <- pdimg_images(c(x1, x2, x3))
res
res[[1]]$path
res[[2]]$path
res[[3]]$path
pdimg_detect(x=res)
}
