% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdimg_detect.R
\name{pdimg_filter}
\alias{pdimg_filter}
\title{pdimg_filter}
\usage{
pdimg_filter(x, min_size = NULL, max_size = NULL)
}
\arguments{
\item{x}{output from \code{\link[=pdimg_images]{pdimg_images()}}}

\item{min_size, max_size}{a file coerceable by \code{\link[fs:fs_bytes]{fs::fs_bytes()}}}
}
\value{
same structure as \code{\link[=pdimg_images]{pdimg_images()}}
}
\description{
filter images by their metadata
}
\examples{
x1 <- system.file("examples/Tierney2017JOSS.pdf", package="pdfimager")
x2 <- system.file("examples/FuHughey2019.pdf", package="pdfimager")
x3 <- system.file("examples/vanGemert2018.pdf", package="pdfimager")
res <- pdimg_images(c(x1, x2, x3))
res
res[[1]]$path
res[[2]]$path
res[[3]]$path
pdimg_filter(x=res)

x4 <- system.file("examples/Wunderlich2020.pdf", package="pdfimager")
res <- pdimg_images(x4)
pdimg_filter(res, min_size = "100K")
}
