% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdimg_images.R
\name{pdimg_images}
\alias{pdimg_images}
\title{pdimg_images}
\usage{
pdimg_images(paths, base_dir = NULL, ...)
}
\arguments{
\item{paths}{(character) path to a pdf, required}

\item{base_dir}{(character) the base path to collect files into. if \code{NULL}
(default), we use a temp directory}

\item{...}{additional params passed on to \code{pdfimages}. See
\code{\link[=pdimg_help]{pdimg_help()}} for docs}
}
\value{
data.frames of metadata on images in the pdf. if the path is
not found or the path is found but no images are found, then a warning
is thrown and a zero row data.frame is returned
}
\description{
extract images from a pdf
}
\examples{
# images found
x <- system.file("examples/BachmanEtal2020.pdf", package="pdfimager")
res <- pdimg_images(x)
res

z <- system.file("examples/Tierney2017JOSS.pdf", package="pdfimager")
pdimg_images(z)

# many at once
pdimg_images(c(x, z))
## pass custom dir to each path
pdimg_images(c(x, z), file.path(tempdir(), "pepperjackcheese"))

# no images found, but there are actually images 
d <- system.file("examples/LahtiEtal2017.pdf", package="pdfimager")
pdimg_images(d)

# no images found, and there really are no images
w <- system.file("examples/White2015.pdf", package="pdfimager")
pdimg_images(w)

# path not found
pdimg_images("foo-bar")
}
