% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfimager-package.R
\docType{package}
\name{pdfimager-package}
\alias{pdfimager-package}
\alias{pdfimager}
\title{pdfimager}
\description{
Extract images from pdfs using poppler
\url{https://poppler.freedesktop.org/}
}
\keyword{package}
