% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfimager-package.R
\docType{package}
\name{pdfimager-package}
\alias{pdfimager}
\alias{pdfimager-package}
\title{pdfimager: Extract images from pdfs}
\description{
Extract images from pdfs using poppler \url{https://poppler.freedesktop.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sckott/pdfimager (source) https://sckott.github.io/pdfimager/ (docs)}
  \item Report bugs at \url{https://github.com/sckott/pdfimager/issues}
}

}
\author{
\strong{Maintainer}: Scott Chamberlain \email{myrmecocystus@gmail.com}

}
\keyword{internal}
