% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdimg_set_path.R
\name{pdimg_set_path}
\alias{pdimg_set_path}
\title{Set path}
\usage{
pdimg_set_path(path = "pdimages")
}
\arguments{
\item{path}{(character) path to pdimages. required}
}
\value{
nothing, \code{cat()} to console
}
\description{
Set the path to your pdfimages
}
\details{
\code{path} is set on an internal (not exported) package environmenzt
variable \code{pdimages_path}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pdimg_set_path("adfadf")
\dontshow{\}) # examplesIf}
}
